fix function prototypes

+++ include/geometrycentral/surface/stripe_patterns.h
@@ -21,14 +21,14 @@ computeStripePattern(IntrinsicGeometryInterface& geometry, const VertexData<doub
 // Extracts the zero (mod 2pi) level set of the function values defined on the corners
 // Returns a list of vertices and edges suitable for rendering (e.g. with Polyscope)
 // (requires access to explicit vertex positions)
-std::tuple<std::vector<Vector3>, std::vector<std::array<size_t, 2>>>
+std::tuple<std::vector<Vector3>, std::vector<std::array<int, 2>>>
 extractPolylinesFromStripePattern(EmbeddedGeometryInterface& geometry, const CornerData<double>& values,
                                   const FaceData<int>& stripeIndices, const FaceData<int>& fieldIndices,
                                   const VertexData<Vector2>& directionField, bool connectOnSingularities);
 
 // Runs both of the above functions (per-corner data computation and polyline extraction)
 // can optionally connect isolines separated by a singularity using a directionField alignment heuristic
-std::tuple<std::vector<Vector3>, std::vector<std::array<size_t, 2>>>
+std::tuple<std::vector<Vector3>, std::vector<std::array<int, 2>>>
 computeStripePatternPolylines(EmbeddedGeometryInterface& geometry, const VertexData<double>& frequencies,
                               const VertexData<Vector2>& directionField, bool connectIsolinesOnSingularities = true);
